fix(hr): the income-tax slab rate was still a double — my earlier fix was incomplete
An earlier pass moved surchargeRatePct, eobiEmployerRatePct and eobiEmployeeRatePct off binary
floating point and routed them through PercentOfPaisa. It missed TaxSlab.ratePct, so the income
tax slab rate — the single largest deduction on a payslip — was still applied as

    Math.round(excess * ratePct / 100.0)

Found by the phase-29 planner while reading the payroll path for its own reasons, and it is
exactly the failure mode that pass existed to remove. Fixing three of four rate fields and
leaving the biggest one is worse than an obvious omission: the file now looked correct.

It has not visibly hurt because the only slab table in existence is one a developer typed as
whole numbers. It starts mattering the moment an accountant enters 11.500 through the tax
configuration screen phase 29 adds — binary floating point cannot represent that exactly, and a
payslip a paisa out is a payslip an employee disputes.

Jackson deserialises JSONB numerics into BigDecimal without configuration, so stored slab rows
need no migration. Test literals use new BigDecimal("11.0"), not new BigDecimal(11.0) — the
double-taking constructor would reintroduce the same imprecision inside the test that is meant
to prove it gone.

Verified: SlabTaxCalculatorTest 3/3, EobiCalculatorTest 7/7, hr-service verify BUILD SUCCESS.

+++ services/hr-service/src/main/java/io/restaurantos/hr/payroll/tax/TaxSlab.java
@@ -1,9 +1,25 @@
 package io.restaurantos.hr.payroll.tax;
 
+import java.math.BigDecimal;
+
 /**
  * One cumulative income-tax bracket. {@code maxPaisa == null} denotes the open-ended top slab.
  * All monetary fields are in paisa (1 PKR = 100 paisa). Slab rows live in {@code tax_config}
  * (JSONB), never hardcoded in Java — this record is just their in-memory shape.
+ *
+ * <p><b>{@code ratePct} is a {@link BigDecimal}, not a double.</b> It was a double, and the earlier
+ * pass that moved {@code surchargeRatePct}, {@code eobiEmployerRatePct} and
+ * {@code eobiEmployeeRatePct} off floating point missed it — so the income-tax slab rate, which is
+ * the single largest deduction on a payslip, was still being applied as
+ * {@code Math.round(excess * ratePct / 100.0)}.
+ *
+ * <p>It has not visibly hurt while the only slab table was one a developer typed as whole numbers.
+ * It starts mattering the moment an accountant enters {@code 11.500} through the tax-configuration
+ * screen that phase 29 adds — binary floating point cannot represent that exactly, and a payslip
+ * that is a paisa out is a payslip an employee disputes.
+ *
+ * <p>Jackson deserialises JSONB numerics into {@code BigDecimal} without configuration, so the
+ * stored slab rows need no migration.
  */
-public record TaxSlab(long minPaisa, Long maxPaisa, long baseTaxPaisa, double ratePct) {
+public record TaxSlab(long minPaisa, Long maxPaisa, long baseTaxPaisa, BigDecimal ratePct) {
 }
